Moved ODD related fucntions to examples/odd.py
